[chore] 잘못된 패키지 구조 개선
