+++ src/main/java/com/MooBoo/MooBoo_Spring/adapter/outbound/external/bookapi/dto/BookInfo.java
@@ -1,4 +1,4 @@
-package com.MooBoo.MooBoo_Spring.adapter.port.outbound.external.dto;
+package com.MooBoo.MooBoo_Spring.adapter.port.outbound.external.bookapi.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
