--- src/main/java/com/MooBoo/MooBoo_Spring/adapter/port/inbound/api/dto/BookSearchResponse.java
@@ -1,26 +1,0 @@
-package com.MooBoo.MooBoo_Spring.adapter.port.inbound.api.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-/**
- * API 응답을 위한 DTO
- */
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class BookSearchResponse {
-    private String isbn13;
-    private String title;
-    private String description;
-    private String author;
-    private String cover;
-    private String publisher;
-    private String pubDate;
-    private String itemPage;
-
-}
-
